split ram into several files
