+++ src/hack_computer/ram/mod.rs
@@ -1,1 +1,5 @@
-pub mod ram;
+pub mod ram8;
+pub mod ram64;
+pub mod ram512;
+pub mod ram4k;
+pub mod ram16k;
